dummy spielzug thinker

+++ main.c
@@ -72,10 +72,7 @@ int main(int argc, char *argv[]) {
     // öffne Konfigurationsdatei und schreibe Werte in hostname, portnumber & gamekindname
     openconfig(filename);
     printf("Hostname: %s \n", _config.hostname);
-    int fd = connect_to_server();
 
-    if(fd == -1)
-        return EXIT_FAILURE;
 
 
 
@@ -88,7 +85,7 @@ int main(int argc, char *argv[]) {
     shmdata->player_number = player_number;
     shmdata->game_name = game_id;
 
-    performConnection(fd, _shm_id);
+   
     
     //Anlegen von namenlosen Pipe
     int feld[2];
@@ -114,15 +111,18 @@ int main(int argc, char *argv[]) {
     else if (pid > 0) {
         printf("Hi hier ist der Thinker (Elternprozess)\n");
         //shmdata->process_id_thinker = pid;
-        
-        signal(SIGUSR1, think);
-        
+    
         //Leseseite schliessen
         close (feld[0]);
         // In die Schreibseite der Pipe schreiben 
-        write (feld[1], puffer, PIPE_BUF);
-
-         ret_code = wait(NULL);
+        //write (feld[1], puffer, PIPE_BUF);
+        
+        fd = feld[1];
+        
+        signal(SIGUSR1, think);
+        
+        
+        ret_code = wait(NULL);
 
         if (ret_code < 0) {
             perror ("Fehler beim Warten auf Kindprozess.");
@@ -138,8 +138,14 @@ int main(int argc, char *argv[]) {
        // shmdata->process_id_connector = pid;
 
        // sleep(10);
+        int _fd = connect_to_server();
+        
+        if(_fd == -1)
+            return EXIT_FAILURE;
        game_state *_game_state = address_shm(_shm_id);
-    
+        performConnection(_fd, _shm_id);
+        
+        
     	if(_game_state->flag_thinking==1){
 			if (kill(getppid(), SIGUSR1) < 0) {
        		perror ("Fehler bei Senden vom Signal).");
